feat(M6): implement real JAX compute path for spatial rate computation
Previously, backend="jax" only converted outputs to JAX arrays at the end.
Now the core math (kernel smoothing, rate computation) uses actual JAX
operations for GPU acceleration and JAX transformation compatibility.

Changes:
- Added backend parameter to smooth_rate_map and smooth_rate_maps_batch
- Added _smooth_rate_map_jax and _smooth_rate_maps_batch_jax implementations
- Modified compute_spatial_rate(s) to pass backend to smoothing functions
- Uses jnp.asarray with explicit dtype=jnp.float64 for precision
- Fixed dtype in view.py and egocentric.py JAX conversions

JAX operations:
- Kernel smoothing: kernel_j @ spike_counts_j (matrix multiply)
- Rate computation: jnp.where for safe division
- Uses _core_jax.compute_firing_rate_single for binned method

Note: JAX defaults to float32 unless JAX_ENABLE_X64=1 is set.
With x64 enabled, results match NumPy to machine precision (~5e-15).

+++ src/neurospatial/encoding/egocentric.py
@@ -1205,8 +1205,8 @@ def compute_egocentric_rate(
     if resolved_backend == "jax" and is_jax_available():
         import jax.numpy as jnp
 
-        firing_rate = jnp.asarray(firing_rate)  # type: ignore[assignment]
-        occupancy = jnp.asarray(occupancy)  # type: ignore[assignment]
+        firing_rate = jnp.asarray(firing_rate, dtype=jnp.float64)
+        occupancy = jnp.asarray(occupancy, dtype=jnp.float64)  # type: ignore[assignment]
 
     # Return result
     return EgocentricRateResult(
@@ -1496,7 +1496,7 @@ def compute_egocentric_rates(
             import jax.numpy as jnp
 
             firing_rates_result = jnp.asarray(firing_rates_result)
-            occupancy = jnp.asarray(occupancy)  # type: ignore[assignment]
+            occupancy = jnp.asarray(occupancy, dtype=jnp.float64)  # type: ignore[assignment]
         return EgocentricRatesResult(
             firing_rates=firing_rates_result,
             occupancy=occupancy,
@@ -1536,8 +1536,8 @@ def compute_egocentric_rates(
     if resolved_backend == "jax" and is_jax_available():
         import jax.numpy as jnp
 
-        firing_rates = jnp.asarray(firing_rates)  # type: ignore[assignment]
-        occupancy = jnp.asarray(occupancy)  # type: ignore[assignment]
+        firing_rates = jnp.asarray(firing_rates, dtype=jnp.float64)
+        occupancy = jnp.asarray(occupancy, dtype=jnp.float64)  # type: ignore[assignment]
 
     # Return result
     return EgocentricRatesResult(
